pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lawnberry-pi"
version = "0.1.0"
description = "LawnBerry Pi v2 Unified System"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0", 
    "httpx>=0.27.0",
    "pydantic>=2.8.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "PyYAML>=6.0.1",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=23.1.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    # ARM64-friendly hardware access libs (optional at runtime; imported lazily)
    "smbus2>=0.4.3",
    "pyserial>=3.5",
    # Geometry validation for map overlap checks (libgeos available via Pi OS repos)
    "shapely>=2.0.4",
    # Numerics for orientation math and latency aggregation without pulling accelerator deps
    "numpy>=1.26.4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
asyncio_mode = "auto"

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E","F","I","UP","B"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"backend/src/api/rest.py" = [
    # Legacy file with mixed import placement and length
    "E402", "F401", "I001", "E501", "UP017", "UP006", "UP045", "B904", "F821"
]

[tool.uv]
# Primary index remains PyPI; add piwheels as extra index to prefer ARM64 wheels on Raspberry Pi.
# uv will honor the lockfile (uv.lock) when using `uv sync --frozen`.
index-url = "https://pypi.org/simple"
extra-index-url = ["https://www.piwheels.org/simple"]
